Make plugin configurable via dotted `.sourcegraph-jetbrains.properties`
Normally configuration files for tools (especially global configuration)
tend to begin with a dot to avoid showing up in file listings and other
tools (unless desired).

This change adds `.sourcegraph-jetbrains.properties` as a candidate path
to check for the global configuration file. Since dotted files are
conventional, the path mentioned in the README has been changed to
suggest the dotted file name.

The original filename is still read as a backup for backwards
compatibility.